split ast into item and query
