+++ prql-compiler/src/ast/mod.rs
@@ -1,16 +1,21 @@
-use std::collections::HashMap;
+/// Abstract syntax tree for PRQL language
+/// 
+/// The central struct here is [Node], that can be of different kinds, described with [item::Item].
 
-use anyhow::{anyhow, bail, Result};
-use enum_as_inner::EnumAsInner;
+use anyhow::{bail, Result};
 use serde::{Deserialize, Serialize};
-use strum::{self, Display};
+use strum::{self};
 
 pub use self::dialect::*;
+pub use self::item::*;
+pub use self::query::*;
 use crate::error::{Error, Reason, Span};
 use crate::utils::*;
 
 pub mod ast_fold;
 pub mod dialect;
+pub mod item;
+pub mod query;
 
 /// A name. Generally columns, tables, functions, variables.
 pub type Ident = String;
@@ -23,201 +28,6 @@ pub struct Node {
     pub span: Option<Span>,
     #[serde(skip)]
     pub declared_at: Option<usize>,
-}
-
-#[derive(Debug, EnumAsInner, Display, PartialEq, Clone, Serialize, Deserialize)]
-pub enum Item {
-    Ident(Ident),
-    String(String),
-    Raw(String),
-    NamedExpr(NamedExpr),
-    Query(Query),
-    /// Generic pipeline that (may) start with a value
-    Pipeline(Pipeline),
-    /// Pipeline operating on a frame. Does not start with a value
-    FramePipeline(Vec<Transform>),
-    Transform(Transform),
-    List(Vec<ListItem>),
-    Range(Range),
-    Expr(Vec<Node>),
-    FuncDef(FuncDef),
-    FuncCall(FuncCall),
-    Table(Table),
-    SString(Vec<InterpolateItem>),
-    FString(Vec<InterpolateItem>),
-    Interval(Interval),
-}
-
-#[derive(Debug, PartialEq, Clone, Serialize, Deserialize, Default)]
-pub struct Query {
-    pub version: Option<String>,
-    #[serde(default)]
-    pub dialect: Dialect,
-    pub nodes: Vec<Node>,
-}
-
-#[derive(Debug, PartialEq, Clone, Serialize, Deserialize)]
-pub struct ListItem(pub Node);
-
-impl ListItem {
-    pub fn into_inner(self) -> Node {
-        self.0
-    }
-}
-
-/// Transformation is used for each stage in a pipeline
-/// and sometimes
-#[derive(Debug, PartialEq, Clone, Serialize, Deserialize)]
-// We probably want to implement some of these as Structs rather than just
-// `vec<Item>`
-pub enum Transform {
-    From(TableRef),
-    Select(Vec<Node>),
-    Filter(Vec<Node>),
-    Derive(Vec<Node>),
-    Aggregate(Vec<Node>),
-    Sort(Vec<ColumnSort<Node>>),
-    Take(i64),
-    Join {
-        side: JoinSide,
-        with: TableRef,
-        filter: JoinFilter,
-    },
-    Group {
-        by: Vec<Node>,
-        pipeline: Vec<Transform>,
-    },
-}
-
-impl Transform {
-    /// Returns the name of the transformation.
-    pub fn name(&self) -> &'static str {
-        match self {
-            Transform::From(_) => "from",
-            Transform::Select(_) => "select",
-            Transform::Filter(_) => "filter",
-            Transform::Derive(_) => "derive",
-            Transform::Aggregate { .. } => "aggregate",
-            Transform::Sort(_) => "sort",
-            Transform::Take(_) => "take",
-            Transform::Join { .. } => "join",
-            Transform::Group { .. } => "group",
-        }
-    }
-
-    pub fn first_node(&self) -> Option<&Node> {
-        match &self {
-            Transform::From(_) => None,
-            Transform::Select(nodes)
-            | Transform::Filter(nodes)
-            | Transform::Derive(nodes)
-            | Transform::Aggregate(nodes) => nodes.first(),
-            Transform::Sort(columns) => columns.first().map(|c| &c.column),
-            Transform::Join { filter, .. } => filter.nodes().first(),
-            Transform::Group { by, .. } => by.first(),
-            Transform::Take(_) => None,
-        }
-    }
-}
-
-#[derive(Debug, PartialEq, Clone, Serialize, Deserialize)]
-pub enum JoinFilter {
-    On(Vec<Node>),
-    Using(Vec<Node>),
-}
-
-impl JoinFilter {
-    fn nodes(&self) -> &Vec<Node> {
-        match self {
-            JoinFilter::On(nodes) => nodes,
-            JoinFilter::Using(nodes) => nodes,
-        }
-    }
-}
-
-/// Function definition.
-#[derive(Debug, PartialEq, Clone, Serialize, Deserialize)]
-pub struct FuncDef {
-    pub name: Ident,
-    pub positional_params: Vec<Node>, // ident
-    pub named_params: Vec<Node>,      // named expr
-    pub body: Box<Node>,
-}
-
-/// Function call.
-#[derive(Debug, PartialEq, Clone, Serialize, Deserialize)]
-pub struct FuncCall {
-    pub name: Ident,
-    pub args: Vec<Node>,
-    pub named_args: HashMap<Ident, Box<Node>>,
-}
-
-#[derive(Debug, PartialEq, Clone, Serialize, Deserialize)]
-pub struct Pipeline {
-    pub value: Option<Box<Node>>,
-    pub functions: Vec<Node>,
-}
-
-#[derive(Debug, PartialEq, Clone, Serialize, Deserialize)]
-pub struct Table {
-    pub name: String,
-    pub pipeline: Vec<Transform>,
-}
-
-#[derive(Debug, PartialEq, Clone, Serialize, Deserialize)]
-pub struct NamedExpr {
-    pub name: Ident,
-    pub expr: Box<Node>,
-}
-
-#[derive(Debug, PartialEq, Clone, Serialize, Deserialize)]
-pub enum InterpolateItem {
-    String(String),
-    Expr(Node),
-}
-
-#[derive(Debug, PartialEq, Clone, Serialize, Deserialize)]
-pub enum JoinSide {
-    Inner,
-    Left,
-    Right,
-    Full,
-}
-
-#[derive(Debug, PartialEq, Clone, Serialize, Deserialize)]
-pub struct TableRef {
-    pub name: String,
-    pub alias: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub struct ColumnSort<T = Node> {
-    pub direction: SortDirection,
-    pub column: T,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub enum SortDirection {
-    Asc,
-    Desc,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub struct Range {
-    pub start: Option<Box<Node>>,
-    pub end: Option<Box<Node>>,
-}
-
-// I could imagine there being a wrapper of this to represent "2 days 3 hours".
-// Or should that be written as `2days + 3hours`?
-//
-// #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-// pub struct Interval(pub Vec<IntervalPart>);
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub struct Interval {
-    pub n: i64,       // Do any DBs use floats or decimals for this?
-    pub unit: String, // Could be an enum IntervalType,
 }
 
 impl Node {
@@ -309,18 +119,5 @@ impl From<Item> for Node {
             span: None,
             declared_at: None,
         }
-    }
-}
-
-impl Default for SortDirection {
-    fn default() -> Self {
-        SortDirection::Asc
-    }
-}
-
-impl From<Item> for anyhow::Error {
-    // https://github.com/bluejekyll/enum-as-inner/issues/84
-    fn from(item: Item) -> Self {
-        anyhow!("Failed to convert {item:?}")
     }
 }
